Simplify Emacs' UI
I was unable to get rid of the welcome screen. There's a conflict with
the way Home Manager adds packages it installs to the .emacs file in my home
directory. Well, that and that I can't get the same configuration I need
in my init.el instead because it is loaded too early and gets overidden
or something. (I also tried moving my init.el to not be injected through
extraConfig, but that didn't work because Emacs opens it read-write and
the Nix store is read-only. Threw warnings.)