Move allowUnfree up to the flake's definition
home-manager disallows the `nixpkgs` options when using the "global"
nixpkgs in the home-manager NixOS module.

> - ken profile: `nixpkgs` options are disabled when `home-manager.useGlobalPkgs` is e>

I had been using those options to set allowUnfree in the standalone
version of home-manager on my work laptop.

I don't want to bring in a second nixpkgs just for standalone
home-manager, so this is an attempt to work around that. (It actually took a
few tries. This GitHub search provided the breakthrough. [0])

[0] https://github.com/search?q=import+nixpkgs+{+config+%3D+{+allowUnfree+%3D+true%3B+}%3B+}%3B+filename%3Aflake.nix&type=Code&ref=advsearch&l=&l=